pyproject.toml
[tool.poetry]
name = "globule"
version = "0.1.0"
description = "A semantic thought processor for capturing and retrieving thoughts intelligently"
authors = ["User <user@example.com>"]
readme = "README.md"
packages = [{include = "globule"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0"
sentence-transformers = "^2.0"
numpy = "^1.24"
aiosqlite = "^0.19"
pydantic = "^2.0"
rich = "^13.0"
python-dotenv = "^1.0"
msgpack = "^1.0"
pyyaml = "^6.0"
httpx = "^0.27"
ollama = "^0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"
black = "^23.0"
isort = "^5.0"

[tool.poetry.scripts]
globule = "globule.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"